Migrating from Scala2 to Scala3

+++ restonomer-core/src/test/scala/com/clairvoyant/restonomer/core/transformation/AddLiteralColumnTransformationSpec.scala
@@ -86,7 +86,7 @@ class AddLiteralColumnTransformationSpec extends CoreSpec with DataFrameMatchers
     val restonomerTransformation = AddLiteralColumn(
       columnName = "col_D",
       columnValue = "1",
-      columnDataType = Some("int")
+      columnDataType = Some("long")
     )
 
     val expectedRestonomerResponseTransformedDF: DataFrame =
